+++ nuxt/stories/card.stories.js
@@ -103,6 +103,7 @@ export const BlogPost = {
     url: 'javascript:void(0);',
     layout: 'post',
     title: 'Can Microsoft Teams call 911?',
+    date: '2021-05-11',
     moment: '2021-05-11',
     author: ['Alexander Rudolph', 'Next Author'],
     categories: ['Microsoft', 'Teams', 'Voice', 'SBC', 'Routing'],
@@ -120,6 +121,7 @@ export const BlogPostLarge = {
     url: 'javascript:void(0);',
     layout: 'post',
     title: 'Azure Virtual Desktop – Ja, aber automatisiert bitte! Lorem ipsum dolor sit',
+    date: '2021-05-11',
     moment: '2021-05-11',
     author: ['Christian Kordel', 'Next Author'],
     categories: ['Microsoft', 'Teams', 'Voice', 'SBC', 'Routing'],
@@ -138,6 +140,7 @@ export const WithAnimation = {
     url: 'javascript:void(0);',
     layout: 'post',
     title: 'Can Microsoft Teams call 911?',
+    date: '2021-05-11',
     moment: '2021-05-11',
     author: ['Alexander Rudolph', 'Next Author'],
     categories: ['Microsoft', 'Teams', 'Voice', 'SBC', 'Routing'],
@@ -153,6 +156,8 @@ export const CardLong = {
   args: {
     long: true,
     scope: 'Umfang: 1-2 Tage',
+    date: '2025-05-11',
+    moment: '2025-05-11',
     layout: 'post',
     title: 'glueckkanja Quick-Check',
     footer:
@@ -183,6 +188,7 @@ export const Webcast = {
   args: {
     youtubeUrl: 'https://www.youtube.com/watch?v=m2pa27Hv8K0',
     title: 'Neue Headline 2. Thema, maximal 2 Zeilig',
+    date: '2022-02-16',
     moment: '2022-02-16',
     author: ['Nadine Kern', 'Next Author'],
     webcast: true,
@@ -246,6 +252,68 @@ export const Product = {
         ],
       },
     ],
+    logo: {
+      img: 'products/scepman/scepman-logo-all-white.svg',
+      cloudinary: true,
+      bgColor: 'var(--color-yellow)',
+    },
+  },
+};
+
+export const ProductWithImage = {
+  args: {
+    img: 'blog/heads/head-emergency-calls.png',
+    cloudinary: true,
+    alt: 'alt text',
+    title: 'Unified Contact',
+    blogtitlepic: 'products/unified-contacts/unified-contacts-products-header.png',
+    tag: {
+      text: 'Free',
+      color: 'white',
+      bgColor: 'var(--color-cream-can)',
+    },
+    product: [
+      {
+        title: 'Tenant',
+        subpoints: [
+          {
+            hideIcon: true,
+            subpoint: 'Lorem Ipsum 1',
+          },
+          {
+            included: true,
+            subpoint: 'Lorem Ipsum 2',
+          },
+        ],
+      },
+      {
+        title: 'Contact',
+        subpoints: [
+          {
+            included: true,
+            subpoint: 'Lorem Ipsum 1',
+          },
+          {
+            included: true,
+            subpoint: 'Lorem Ipsu 2m',
+          },
+          {
+            subpoint: 'Lorem Ipsum 3',
+          },
+        ],
+      },
+      {
+        title: 'Other',
+        subpoints: [
+          {
+            subpoint: 'Lorem Ipsum 1',
+          },
+          {
+            subpoint: 'Lorem Ipsum 2',
+          },
+        ],
+      },
+    ],
   },
 };
 
@@ -255,6 +323,7 @@ export const Row = {
     url: 'javascript:void(0);',
     layout: 'post',
     title: 'Can Microsoft Teams call 911?',
+    date: '2021-05-11',
     moment: '2021-05-11',
     author: ['Alexander Rudolph', 'Next Author'],
     categories: ['Microsoft', 'Teams', 'Voice', 'SBC', 'Routing'],